♻️ move Encoder and createIdentityEncoder to @datadog/js-core/transport
